+++ src/components/PlayArea/style.tsx
@@ -1,4 +1,5 @@
 import styled from 'styled-components';
+import Breakpoints from '@utilities/breakpoints';
 
 const Style = styled.div`
   border: 1px solid blue;
@@ -7,6 +8,16 @@ const Style = styled.div`
   grid-template-columns: repeat(10, 156px [col-start]);
   margin: 0 auto;
   width: 1560px;
+
+  ${Breakpoints.Medium} {
+    grid-template-columns: repeat(8, 156px [col-start]);
+    width: 1248px;
+  }
+
+  ${Breakpoints.Small} {
+    grid-template-columns: repeat(6, 156px [col-start]);
+    width: 936px;
+  }
 `;
 
 export default Style;