+++ docs/changelog.md
@@ -84,3 +84,4 @@ Ended August 3, 2020
 - add mousedown logic to Card component to toggle the `.active` class.
 - create utlities folder.
 - add Breakpoints enums.
+- add small and medium breakpoints to PlayArea component.